+++ CHANGELOG.md
@@ -7,6 +7,7 @@ using the [Keep a CHANGELOG](http://keepachangelog.com) principles.
 ## [Unreleased]
 ### Added
   - Add PSR2 standard
-  - Add rule to fix wrong variable identation
+  - Add rule to fix wrong code identation
+  - Add rule to fix line and block comment identation
   - Add spacing between class properties
   - Add spacing between functions